requirements.txt
apache-airflow==3.1.0
apache-airflow-providers-google==10.16.0
google-cloud-bigquery
google-cloud-storage
